+++ AtlasLootClassic_DungeonsAndRaids/data.lua
@@ -952,6 +952,7 @@ data["Gnomeregan"] = {
 				{ 7,  9488 }, -- Oscillating Power Hammer
 				{ 8,  9486 }, -- Supercharger Battle Axe
 				{ 9,  9490 }, -- Gizmotron Megachopper
+				{ 11, 9489 }, -- Gyromatic Icemaker
 				{ 16, 9327 }, -- Security DELTA Data Access Card
 				{ 18, 7191 }, -- Fused Wiring
 				{ 19, 9308 }, -- Grime-Encrusted Object
@@ -5753,17 +5754,18 @@ data["Zul'Gurub"] = {
 				{ 2,  20257, [ATLASLOOT_IT_ALLIANCE] = 20264 }, -- Seafury Gauntlets
 				--{ 3,  20264, [ATLASLOOT_IT_HORDE] = 20257 }, -- Peacekeeper Gauntlets
 				{ 3,  19855 }, -- Bloodsoaked Legplates
-				{ 6,  19876 }, -- Soul Corrupter's Necklace
-				{ 7,  19856 }, -- The Eye of Hakkar
-				{ 8,  19861 }, -- Touch of Chaos
-				{ 9,  19853 }, -- Gurubashi Dwarf Destroyer
-				{ 10, 19862 }, -- Aegis of the Blood God
-				{ 11, 19864 }, -- Bloodcaller
-				{ 12, 19865 }, -- Warblade of the Hakkari
-				{ 13, 19852 }, -- Ancient Hakkari Manslayer
-				{ 14, 19859 }, -- Fang of the Faceless
-				{ 15, 19854 }, -- Zin'rokh, Destroyer of Worlds
-				{ 16, 19802 }, -- Heart of Hakkar
+				{ 4,  19876 }, -- Soul Corrupter's Necklace
+				{ 5,  19856 }, -- The Eye of Hakkar
+				{ 7, 19802 }, -- Heart of Hakkar
+				{ 16,  19861 }, -- Touch of Chaos
+				{ 17,  19853 }, -- Gurubashi Dwarf Destroyer
+				{ 18, 19862 }, -- Aegis of the Blood God
+				{ 19, 19864 }, -- Bloodcaller
+				{ 20, 19865 }, -- Warblade of the Hakkari
+				{ 21, 19866 }, -- Warblade of the Hakkari
+				{ 22, 19852 }, -- Ancient Hakkari Manslayer
+				{ 23, 19859 }, -- Fang of the Faceless
+				{ 24, 19854 }, -- Zin'rokh, Destroyer of Worlds
 			},
 		},
 		{ -- ZGShared
@@ -6177,6 +6179,10 @@ data["TheRuinsofAhnQiraj"] = { -- AQ20
 				{ 5,  21494 }, -- Southwind's Grasp
 				{ 6,  21495 }, -- Legplates of the Qiraji Command
 				{ 7,  21497 }, -- Boots of the Qiraji General
+				{ 9,  "INV_Box_01", nil, AL["Trash"] },
+				{ 10,  21810 }, -- Treads of the Wandering Nomad
+				{ 11,  21809 }, -- Fury of the Forgotten Swarm
+				{ 12,  21806 }, -- Gavel of Qiraji Authority
 				{ 19, 20885 }, -- Qiraji Martial Drape
 				{ 20, 20889 }, -- Qiraji Regal Drape
 				{ 21, 20888 }, -- Qiraji Ceremonial Ring