+++ AtlasLootClassic/Documentation/Release_Notes.md
@@ -2,6 +2,10 @@
 
 This mod is distributed under Version 2 of the GPL.  A copy of the GPL is included in this zip file with links to non-english translations.
 
+## v1.5.4 (Jul. 31, 2020)
+
+- update loot sources for Phase 5
+
 ## v1.5.3 (Jul. 30, 2020)
 
 - Update content phase